pyproject.toml
[project]
name = "active-rlhf"
version = "0.1.0"
description = ""
authors = [
    {name = "Benjamin Kuen",email = "B.Kuen@campus.lmu.de"}
]
readme = "README.md"
requires-python = "~3.10"
dependencies = [
    "gymnasium[mujoco] (>=1.1.1,<2.0.0)",
    "wandb (>=0.19,<0.20)",
    "tyro (>=0.9.5,<0.10.0)",
    "torch (>=2.5,<2.6)",
    "numpy (>=2.0.0,<2.1.0)",
    "tqdm (>=4.67,<4.68)",
    "tensorboard (>=2.18,<2.19)",
    "stable-baselines3 (>=2.6,<2.7)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "umap-learn (>=0.5.7,<0.6.0)",
]

packages = [{include = "active_rlhf", from="src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
